my_trading_bot/requirements.txt
# requirements.txt
ccxt>=4.2.0
pandas>=1.5.0
python-dotenv>=1.0.0